# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "http://80.78.246.59/Stech/public/api/CreateTender": {
            "post": {
                "tags": [
                    "tender"
                ],
                "summary": "CreateTender",
                "description": "CreateTender",
                "operationId": "CreateTender",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "string",
                                        "format": "category_id",
                                        "example": "1"
                                    },
                                    "category_name": {
                                        "type": "string",
                                        "format": "category_name",
                                        "example": "Грузо-перевозки"
                                    },
                                    "sub_category_id": {
                                        "type": "string",
                                        "format": "sub_category_id",
                                        "example": "2"
                                    },
                                    "sub_category_name": {
                                        "type": "string",
                                        "format": "sub_category_name",
                                        "example": "2"
                                    },
                                    "region_id": {
                                        "type": "string",
                                        "format": "region_id",
                                        "example": "1"
                                    },
                                    "region_name": {
                                        "type": "string",
                                        "format": "region_name",
                                        "example": "Москва и Московская обл."
                                    },
                                    "city_id": {
                                        "type": "string",
                                        "format": "city_id",
                                        "example": "1"
                                    },
                                    "city_name": {
                                        "type": "string",
                                        "format": "city_name",
                                        "example": "Москва"
                                    },
                                    "street": {
                                        "type": "string",
                                        "format": "street",
                                        "example": "Абрамцевская улица"
                                    },
                                    "date_time": {
                                        "type": "string",
                                        "format": "date_time",
                                        "example": "13/08/2022 09:00"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dumm"
                                    },
                                    "photo[]": {
                                        "type": "file",
                                        "format": "photo[]",
                                        "example": "photo.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/getMytender": {
            "get": {
                "tags": [
                    "tender"
                ],
                "summary": "Get list of Category",
                "description": "Returns list of Category",
                "operationId": "getMytender",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/deleteTender": {
            "post": {
                "tags": [
                    "tender"
                ],
                "summary": "deleteTender",
                "description": "deleteTender",
                "operationId": "deleteTender",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tender_id"
                                ],
                                "properties": {
                                    "tender_id": {
                                        "type": "string",
                                        "format": "tender_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/DeleteTenderPhoto": {
            "post": {
                "tags": [
                    "tender"
                ],
                "summary": "DeleteTenderPhoto",
                "description": "DeleteTenderPhoto",
                "operationId": "DeleteTenderPhoto",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tender_id"
                                ],
                                "properties": {
                                    "tender_id": {
                                        "type": "string",
                                        "format": "tender_id",
                                        "example": "1"
                                    },
                                    "photo_id": {
                                        "type": "string",
                                        "format": "photo_id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/UpdateTender": {
            "post": {
                "tags": [
                    "tender"
                ],
                "summary": "UpdateTender",
                "description": "UpdateTender",
                "operationId": "UpdateTender",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tender_id"
                                ],
                                "properties": {
                                    "tender_id": {
                                        "type": "string",
                                        "format": "tender_id",
                                        "example": "1"
                                    },
                                    "photo[]": {
                                        "type": "file",
                                        "format": "photo[]",
                                        "example": "photo.png"
                                    },
                                    "region_id": {
                                        "type": "string",
                                        "format": "region_id",
                                        "example": "1"
                                    },
                                    "region_name": {
                                        "type": "string",
                                        "format": "region_name",
                                        "example": "Москва и Московская обл."
                                    },
                                    "city_id": {
                                        "type": "string",
                                        "format": "city_id",
                                        "example": "1"
                                    },
                                    "city_name": {
                                        "type": "string",
                                        "format": "city_name",
                                        "example": "Москва"
                                    },
                                    "street": {
                                        "type": "string",
                                        "format": "street",
                                        "example": "Ivanovo 45/1"
                                    },
                                    "date_time": {
                                        "type": "string",
                                        "format": "date_time",
                                        "example": "13/08/2022 09:00"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description",
                                        "example": "description   min:30 charset"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "http://80.78.246.59/Stech/public/api/getAllTenderForRoleId3": {
            "get": {
                "tags": [
                    "tender"
                ],
                "summary": "Get list of getAllTenderForRoleId3",
                "description": "Returns list of getAllTenderForRoleId3",
                "operationId": "getAllTenderForRoleId3",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}